src/ recommendation from framework
It has been decided that it is better to separate charm source from
libraries such as the framework, even though that means that you have to
modify `sys.path` prior to import. This pattern is likely to evolve as
we determine the cleanest way to integrate with the framework while
keeping dependencies separated from charm code.

This also includes a few small tweaks to fix lint errors and remove
dot files mistakenly added from the framework.